requirements.txt
google-generativeai
pygithub
python-dotenv
